+++ src/tetra_rp/core/resources/resource_manager.py
@@ -1,14 +1,14 @@
 import cloudpickle
+import logging
 from typing import Dict
 from pathlib import Path
 
-from tetra_rp import get_logger
-from tetra_rp.core.utils.singleton import SingletonMixin
+from ..utils.singleton import SingletonMixin
 
 from .base import DeployableResource
 
 
-log = get_logger("resource_manager")
+log = logging.getLogger(__name__)
 
 # File to persist state of resources
 RESOURCE_STATE_FILE = Path(".tetra_resources.pkl")
@@ -69,9 +69,11 @@ class ResourceManager(SingletonMixin):
                 return await self.get_or_deploy_resource(config)
 
             log.debug(f"{existing} exists, reusing.")
+            log.info(f"URL: {existing.url}")
             return existing
 
         if deployed_resource := await config.deploy():
+            log.info(f"URL: {deployed_resource.url}")
             self.add_resource(uid, deployed_resource)
             return deployed_resource